requirements.txt
fastapi
uvicorn
Pillow
google-genai
mangum